[TH-1152] 홈 하위 이슈 수정 반영

+++ app/src/main/java/com/zion830/threedollars/ui/storeDetail/boss/viewModel/BossStoreDetailViewModel.kt
@@ -244,7 +244,7 @@ class BossStoreDetailViewModel @Inject constructor(
         }
     }
 
-    fun putStoreReview(reviewId: Int, content: String, rating: Int) {
+    fun putStoreReview(reviewId: Long, content: String, rating: Int) {
         viewModelScope.launch(coroutineExceptionHandler) {
             homeRepository.putStoreReview(reviewId, content, rating).collect {
                 if (it.ok) {
@@ -272,7 +272,7 @@ class BossStoreDetailViewModel @Inject constructor(
         }
     }
 
-    fun reportReview(storeId: Int, reviewId: Int, reportReviewModelRequest: ReportReviewModelRequest) {
+    fun reportReview(storeId: Int, reviewId: Long, reportReviewModelRequest: ReportReviewModelRequest) {
         viewModelScope.launch {
             homeRepository.reportStoreReview(storeId, reviewId, reportReviewModelRequest).collect {
                 if (it.ok) {
